+++ source/calculators/fpfn.cpp
@@ -9,7 +9,7 @@
 #include "fpfn.hpp"
 
 /***********************************************************************/
-double FPFN::getValue( long long tp,  long long tn,  long long fp,  long long fn) {
+double FPFN::getValue(double tp,  double tn,  double fp,  double fn) {
     try {
         long long p = fp + fn;
